Navbar and path
update three diff navbars and fix paths
