+++ frontend/src/components/HealthDataAnalyzer.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import axios from 'axios';
 import { ReactTyped as Typed } from 'react-typed';
 import styled, { createGlobalStyle } from 'styled-components';
-import NavigationBar from './NavigationBar';
+import NavigationBar from './Navbar_master';
 
 const GlobalStyle = createGlobalStyle`
   body {
